Move Frontend and FrontendState to separate modules
